+++ include/mdu/mdu.hpp
@@ -15,3 +15,5 @@
 #include "rx/zsu_base.hpp"
 #include "rx/zsu_zpp_base.hpp"
 #include "tx/base.hpp"
+#include "tx/packets_base.hpp"
+#include "tx/timings_base.hpp"