+++ include/mdu/tx/timings_base.hpp
@@ -1,0 +1,19 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+/// Transmit timings base
+///
+/// \file   mdu/tx/timings_base.hpp
+/// \author Vincent Hamp
+/// \date   17/09/2026
+
+#pragma once
+
+#include "base.hpp"
+
+namespace mdu::tx {
+
+using TimingsBase = Base<Timings>;
+
+} // namespace mdu::tx