+++ tests/tx/mock.hpp
@@ -2,9 +2,9 @@
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
-#include <mdu/tx/base.hpp>
+#include <mdu/mdu.hpp>
 
-struct Mock : mdu::tx::Base<Mock, mdu::Packet> {
+struct Mock : mdu::tx::PacketsBase {
   MOCK_METHOD(void, ackreqBegin, ());
   MOCK_METHOD(void, ackreqChannel1, (size_t));
   MOCK_METHOD(void, ackreqChannel2, (size_t));